remove misleading comment
relates to #852 .  Hi @Affie, think you misunderstanding whats happening here.  The design decision we all made was to keep DFGFactor immutable, but that means extracting a subgraph with only a list of variables must avoid orphaned factors -- the warning true by default is good enough to resolve that issue.  Lets please move on from this discussion, the subgraph consolidation is finalized for the foreseeable future.